Change right-click-drag to control zoom a la MC5D
Right-click-drag results in manually adjusting the 4D viewpoint.
However, the ctrl+left-click functionality of the program is already good
for snapping the puzzles to specific viewpoints that can help with solves.

Additionally, other sim programs like MC5D use right-click-drag to adjust
the view, so this commit brings in that behaviour, while still keeping the
mousewheel code in case one still wants to use the mousewheel to zoom.